==> 399 - Data Science/3_gini.ipynb <==
<h1>
<center>
Module 3: Gini assignment
</center>
</h1>
<div class=h1_cell>

You will be working with the loan table again.

</div>

import pandas as pd

from google.colab import drive
drive.mount('/content/gdrive')

with open('/content/gdrive/My Drive/Colab Notebooks/399/notebook_history/loan_wrangled_week2.csv', 'r') as f:
  loan_table = pd.read_csv(f)

We changed the library so need to update it for this notebook.

from google.colab import files
files.upload()

#load the lbirary for this week to use what was defined in content notebook

from library_w19_week3 import *
%who function

pd.set_option('display.max_columns', None)
loan_table.head()

<hr>
<h1>
1: One hot encode ApplicantIncome
</h1>
<p>
<div class=h1_cell>
First bin it and then one hot encode the binned values. Use ['low', 'average', 'high'] as the 3 binned values.
</div>

#bin ApplicantIncome
bins = 3
bin_names = ['low', 'average', 'high']
loan_table['apin_binned'] = pd.cut(loan_table['ApplicantIncome'], bins, labels=bin_names)

loan_table.head()

#ohe apin_binned
one_hot_apin = pd.get_dummies(loan_table['apin_binned'], prefix='apin', dummy_na=True)
loan_table = loan_table.join(one_hot_apin)

loan_table.head(1)

<hr>
<h1>
2: Use `ch_bad` as root splitter
</h1>
<p>
<div class=h1_cell>
Build the 2 sub-tables from splitting with `ch_bad` on the root node.
</div>

true_bad = loan_table.loc[loan_table['ch_bad'] == 1]
false_bad = loan_table.loc[loan_table['ch_bad'] == 0]

print((len(true_bad), len(false_bad)))

<hr>
<h1>
3: Determine the best splitter for the true_bad table
</h1>
<p>
<div class=h1_cell>
Using the `true_bad` table as the root, compute a sorted list of the `apin_*` columns based on their gig scores.
</div>

columns = ['apin_average', 'apin_low', 'apin_high', 'apin_nan']

gig_scores = map(lambda col: (col, gig(true_bad, col, 'Loan_Status')), columns)
gig_sorted = sorted(gig_scores, key=lambda item: item[1], reverse=True)

gig_sorted

<hr>
<h1>
4: Explore Entropy
</h1>
<p>
<div class=h1_cell>
<p>
Entropy is an alternative measure to Gini. Here is the formula you need, where p(a) is probablity of Loan_Status being 1 and p(b) is probability of Loan_Status being 0. Also, log is in base 2.
<p>
`Entropy = - p(a)*log(p(a)) - p(b)*log(p(b))`
<p>
Go ahead and define a function `entropy` that takes in a Series object with counts (which I'll give you). It returns the entropy score as above.
<p>
Hint: the `apin_nan` splitter will give you grief if you are not careful. It will produce an empty table on one branch. Be careful to avoid dividing by zero or taking the log of 0.
</div>

#first get probabilities for entire table
root_counts = loan_table['Loan_Status'].value_counts()

root_counts  #422 successful applications

# watch out for 0 valued probs

import math

def entropy(counts):
  (p0, p1) = probabilities(counts)
  left = 0 if p0 == 0 else -(p0 * math.log(p0, 2))
  right = 0 if p1 == 0 else (p1 * math.log(p1, 2))
  return left - right

#compute entropy of entire loan_table before any splits

entropy(root_counts)

<div class=just_text>
Like Gini, to use Entropy for a split on column C, compute the Entropy of the 3 tables, i.e., the starting table, the true table (column C is true) and the false table (column C is false). We now have 3 separate Entropy values for the 3 tables involved. Our next step is to combine these values in a way  that gives us an overall goodness-of-the-split score. To do so, we will use something called *`information gain`* (IGain). The formula for IGain is as follows:
<p>
`IGain = start_entropy − (w_true * entropy_true + w_false * entropy_false)`
<p>
where w_true is a weight of |true_table|/|titanic_table| (i.e., the size of true_table divided by the size of that starting table) and w_false is a weight of |false_table|/|titanic_table|.
<p>
Notice anything? Igain is equivalent to GIG at this point. It combines its 3 scores in the same way. It is just the indiviual scoring method that differs.
<p>
Please define an IGain function and use it to calculate a split on loan_table. Compare it with the gig score on the same split.
</div>

def IGain(starting_table, split_column, target_column):
  true_table = starting_table.loc[starting_table[split_column] == 1]
  false_table = starting_table.loc[starting_table[split_column] == 0]

  true_counts = true_table[target_column].value_counts()
  false_counts = false_table[target_column].value_counts()
  starting_counts = starting_table[target_column].value_counts() 

  #compute the entropy for the 3 tables
  starting_entropy = entropy(starting_counts)
  true_entropy = entropy(true_counts)
  false_entropy = entropy(false_counts)

  #compute the weights
  starting_size = len(starting_table.index)
  true_weight = 0.0 if starting_size == 0 else len(true_table.index)/starting_size
  false_weight = 0.0 if starting_size == 0 else len(false_table.index)/starting_size

  #wrap it up and put on a bow
  IGain = starting_entropy - (true_weight * true_entropy + false_weight * false_entropy)

  return IGain

#gig scores computed here so we can compare


gig_sorted

#igain scores computed here
columns = ['apin_average', 'apin_low', 'apin_high', 'apin_nan']

igain_scores = map(lambda col: (col, IGain(loan_table, col, 'Loan_Status')), columns)
igain_sorted = sorted(igain_scores, key=lambda item: item[1], reverse=True)

igain_sorted

<div class=just_text>
There is a difference but I conjecture that the scores are so low that none of these look like good choices as a splitter. So the difference is in the noise. There are a lot of other columns to explore but I will hold off for now.
  <p>
    
</div>

<hr>
<h1>Write it out</h1>
<div class=h1_cell>

We made changes so save the table so can use it in next module.
</div>

loan_table.head(1)

with open('/content/gdrive/My Drive/Colab Notebooks/399/notebook_history/loan_wrangled_week3.csv', 'w') as f:
  loan_table.to_csv(f, encoding='utf-8', index=False)

<hr>
<h1>No updates for library</h1>
<div class=h1_cell>

We defined new  entropy functions but we won't be using them later - we will stick with gini. So you should be good to go with library_w19_week3.py.
</div>
