==> 399 - Data Science/4_tree_building.ipynb <==
<h1>
<center>
Module 4: Tree Building assignment
</center>
</h1>
<div class=h1_cell>

You will be working with the loan table again.

</div>

import pandas as pd

from google.colab import drive
drive.mount('/content/gdrive')


with open('/content/gdrive/My Drive/Colab Notebooks/399/notebook_history/loan_wrangled_week3.csv', 'r') as f:
  loan_table = pd.read_csv(f)

with open('/content/gdrive/My Drive/Colab Notebooks/399/notebook_history/loan_results_week2.csv', 'r') as f:
  loan_results = pd.read_csv(f)

!rm library_w19_week4.py

from google.colab import files
files.upload()

from library_w19_week4 import *
%who function

pd.set_option('display.max_columns', None)
loan_table.head()

loan_table.columns.values

<hr>
<h1>
1: OHE Dependents
</h1>
If you take a look at the Dependents column below, you will see that it is viewed as a categorical column with 4 unique values. So we should be able to use get_dummies on it.

loan_table.describe(include='all')

Go ahead and ohe the `Dependents` column - use `dep` as prefix. And look at above to see the count. It has empties, right? So set the appropriate flag.

#ohe code here
one_hot_dep = pd.get_dummies(loan_table['Dependents'], prefix='dep', dummy_na=True)
loan_table = loan_table.join(one_hot_dep)

loan_table.head(1)

<hr>
<h1>
2: Find the best depth
</h1>
<p>
<div class=h1_cell>
<p>
Explore depths from 1 to 5. Build a tree for each depth and test it. Plot the results.
<p>
First, define the columns to use. I do that for you below.
</div>

column_candidates = [
        #Dependents
        'dep_0', 'dep_1', 'dep_2', 'dep_3+',
        #ApplicantIncome
       'apin_low', 'apin_high', 'apin_average',
        #Property_Area
        'pa_Rural', 'pa_Semiurban','pa_Urban',
        #LoanAmount
        'lam_Low', 'lam_Average', 'lam_High',
        #Credit_History
        'ch_bad', 'ch_good']

tree1 = build_tree_iter(loan_table, column_candidates, 'Loan_Status', {'max-depth': 1}) # should get 2 paths
print(len(tree1['paths']))
tree1['paths']

<div class=h1_cell>
<p>
As reminder, a tree with only one splitter node is called a stump. You can see that `ch_bad` is the best splitter with only one node. Thought question: As we build deeper trees below, is it possible for the root node to be something other than `ch_bad`?
</div>

tree2 = build_tree_iter(loan_table, column_candidates, 'Loan_Status', {'max-depth': 2}) # should get 4 paths
print(len(tree2['paths']))
tree2['paths']

tree3 = build_tree_iter(loan_table, column_candidates, 'Loan_Status', {'max-depth': 3}) # should get 7 paths
print(len(tree3['paths']))
tree3['paths']

tree4 = build_tree_iter(loan_table, column_candidates, 'Loan_Status', {'max-depth': 4}) # should get 13 paths
print(len(tree4['paths']))
#tree4['paths']

tree5 = build_tree_iter(loan_table, column_candidates, 'Loan_Status', {'max-depth': 5}) # should get 20 paths
print(len(tree5['paths']))
#tree5['paths']

<div class=h1_cell>
  <h1>You have your 5 trees</h1>
<p>
Add them to your results table.
</div>

loan_results['tree_1'] = loan_table.apply(lambda row: tree_predictor(row, tree1), axis=1)

loan_results['tree_1_type'] = loan_results.apply(lambda row: predictor_case(row, pred='tree_1', target='Loan_Status'), axis=1)

loan_results.head(1)

loan_results['tree_2'] = loan_table.apply(lambda row: tree_predictor(row, tree2), axis=1)

loan_results['tree_2_type'] = loan_results.apply(lambda row: predictor_case(row, pred='tree_2', target='Loan_Status'), axis=1)

loan_results.head(1)

loan_results['tree_3'] = loan_table.apply(lambda row: tree_predictor(row, tree3), axis=1)

loan_results['tree_3_type'] = loan_results.apply(lambda row: predictor_case(row, pred='tree_3', target='Loan_Status'), axis=1)

loan_results.head(1)

loan_results['tree_4'] = loan_table.apply(lambda row: tree_predictor(row, tree4), axis=1)

loan_results['tree_4_type'] = loan_results.apply(lambda row: predictor_case(row, pred='tree_4', target='Loan_Status'), axis=1)

loan_results.head(1)

loan_results['tree_5'] = loan_table.apply(lambda row: tree_predictor(row, tree5), axis=1)

loan_results['tree_5_type'] = loan_results.apply(lambda row: predictor_case(row, pred='tree_5', target='Loan_Status'), axis=1)

loan_results.head(1)

Get the value counts on each tree.

tree_1_types = loan_results['tree_1_type'].value_counts()
tree_2_types = loan_results['tree_2_type'].value_counts()
tree_3_types = loan_results['tree_3_type'].value_counts()
tree_4_types = loan_results['tree_4_type'].value_counts()
tree_5_types = loan_results['tree_5_type'].value_counts()

ok, ready to check out our measures.

print((accuracy(tree_1_types), f1(tree_1_types), informedness(tree_1_types)))

print((accuracy(tree_2_types), f1(tree_2_types), informedness(tree_2_types)))

print((accuracy(tree_3_types), f1(tree_3_types), informedness(tree_3_types)))

print((accuracy(tree_4_types), f1(tree_4_types), informedness(tree_4_types)))

print((accuracy(tree_5_types), f1(tree_5_types), informedness(tree_5_types)))

<div class=h1_cell>
<p>
Should we stop at 5? It looks like our measures are continuing to go up. I'll let you stop here but in practice it would be good to find where depth is optimal.
  <p>
Let's try at least one change to the gig cutoff. Use `0.01` and level 5.
</div>

tree6 = build_tree_iter(loan_table, column_candidates, 'Loan_Status', {'max-depth': 5, 'gig-cutoff': .01})

print(len(tree6['paths']))
#tree6['paths']

Wow. We went from 20 paths to 3 by using gig-cutoff. See how tree6 does.

loan_results['tree_6'] = loan_table.apply(lambda row: tree_predictor(row, tree6), axis=1)

loan_results['tree_6_type'] = loan_results.apply(lambda row: predictor_case(row, pred='tree_6', target='Loan_Status'), axis=1)

loan_results.head(1)

tree_6_types = loan_results['tree_6_type'].value_counts()
print((accuracy(tree_6_types), f1(tree_6_types), informedness(tree_6_types)))

Same as tree3 but with 4 fewer paths: 3 instead of 7.

<hr>
<h1>
3: Order the paths in a tree according to most used
</h1>
<p>
<div class=h1_cell>
<p>
One of the cool things about building our own code for doing predictions with decision trees is that we can explore various options directly. Let's do that now. Look at this code in `tree_predictor`:
<p>
```
for path in tree['paths']:
```
<p>
It is doing a sequential and ordered enumeration of the paths in a tree. We can gain some efficiency by ordering the paths, right? If we can determine which paths are most likely to succeed, we can try those paths first. In practice, this means reordeing the list of dictionaries in tree['paths'] so most used path is first, and least used path is last. I'd like you to write some code that will reorder `tree['paths']` on this basis of most used.
<p>
I'll break it up into pieces. First, define a function `path_id` that accepts a row and a tree and returns the index of the path that succeeded. So if the first path in tree['paths'] succeeded, you would return its index 0.
</div>

def path_id(row, tree):
  
    assert (len(tree['paths']) > 0), "Tree must have at least one path"
    
    for idx,path in enumerate(tree['paths']):
      conjuncts = path['conjunction']
      result = map(lambda tuple: tuple[1](row), conjuncts)  # potential to be parallelized
      if all(result):
        return idx

<div class=h1_cell>

Now apply the function against the loan table using tree1 and produce a series called `pcounts1`. Remember that tree1 has only 2 paths.
</div>

pcounts1 = loan_table.apply(lambda row: path_id(row, tree1), axis=1)

pcounts1.value_counts()

<div class=h1_cell>
Looks like path 1 (i.e., `tree1['paths'][1]`) is the big winner.
  <p>
    Let's see what tree2 looks like. It has 4 paths.
</div>

pcounts2 = loan_table.apply(lambda row: path_id(row, tree2), axis=1)

pcounts2.value_counts()

<div class=h1_cell>
  Looks like path 0 should definitely be tried last. In fact, it looks like the list is in exactly reverse order in terms of path usage.
  <p>
Let's look at tree3.
</div>

pcounts3 = loan_table.apply(lambda row: path_id(row, tree3), axis=1)

pcounts3.value_counts()

<div class=h1_cell>

Let's use `tree3` (i.e., `pcounts3`) as our focus. Turn `pcounts3.value_counts()` into a dictionary.
</div>

pdict3 = dict(pcounts3.value_counts())
pdict3

<div class=h1_cell>

Now turn the dictionary into a list of tuples. Be careful. Generators can bite you.
</div>

plist3 = [(key,pdict3[key]) for key in pdict3]
plist3

<div class=h1_cell>
This is kind of odd. It appears that converting to a list of tuples has also sorted the list by value in reverse order. I cannot find any documentation that says this should happen. So I am not taking chances. I am going to go ahead and sort `plist3` in a separate function. You shoud too unless you can point me to documentation that says pulling tuples from a dictionary does sorting.
</div>

plist3 = sorted(plist3, key = lambda tup: tup[1])
plist3 = plist3[::-1]
plist3

<div class=h1_cell>
Finally, build a new path list using plist3 as a guide.
</div>

new_paths = []
for tup in plist3:
  new_paths.append(tree3['paths'][tup[0]])

new_paths

I can now reset `tree3["paths"]` to `new_paths`. Cool.

<hr>
<h2>
Put it all together in a function
</h2>
<p>
<div class=h1_cell>
<p>
Please write `reorder_paths` to produce a list of paths that is sorted on most frequently used. You can use the output of the function as:
<p>
<pre>
    ordered_paths3 = reorder_paths(loan_table, tree3)
    tree3['paths'] = ordered_paths3
</pre>
<p>
I'll give you some examples below that you can test on - match my output. Note that I am printing some debugging info from inside my function. I am printing the sorted list I come up with. You do not have to print this but it is useful for debugging.
</div>

def reorder_paths(table, tree):
  pcounts = table.apply(lambda row: path_id(row, tree), axis = 1)
  pdict = dict(pcounts.value_counts())
  plist = [(key, pdict[key]) for key in pdict]
  plist = sorted(plist, key = lambda tup: tup[1])[::-1]
  
  new_paths = []
  for tup in plist:
    new_paths.append(tree['paths'][tup[0]])
    
  return new_paths

tree1['paths']  #before we reorder

reorder_paths(loan_table, tree1)  # First line printed is just for debugging

tree2['paths']  #before we reorder

reorder_paths(loan_table, tree2)

reorder_paths(loan_table, tree3)

<hr>
<h1>Write it out</h1>
<div class=h1_cell>

We changed both loan_table and loan results. Write them out.
</div>

with open('/content/gdrive/My Drive/Colab Notebooks/399/notebook_history/loan_table_week4.csv', 'w') as f:
  loan_table.to_csv(f, encoding='utf-8', index=False)
  
with open('/content/gdrive/My Drive/Colab Notebooks/399/notebook_history/loan_results_week4.csv', 'w') as f:
  loan_results.to_csv(f, encoding='utf-8', index=False)

<hr>
<h1>Don't forget library functions</h1>
<div class=h1_cell>

Add them to `library_w19_week4b.py`. You defined `library_w19_week4.py` in intro module. You are now defining a new version of it with functions from this notebook. I added the `b` to keep them clear.
</div>
