==> 399 - Data Science/5_kfold.ipynb <==
<h1>
<center>
Module 5: Bias-Variance and Cross-Validation
</center>
</h1>
<div class=h1_cell>

You will be working with the loan table again.

</div>

import pandas as pd

from google.colab import drive
drive.mount('/content/gdrive')

with open('/content/gdrive/My Drive/class_tables/loan_wrangled_week3.csv', 'r') as f:
  loan_table = pd.read_csv(f)

loan_table.head(2)  #make sure it looks ok - we see the results of our week 2 wrangling

pd.set_option('display.max_columns', None)

<h2>A couple of notes</h2>
I added points to assignments to give you an idea of how much of a challenge I think they are. Let me know if I am off base.
<p>
  Don't load your library yet. Finish problem 1 before importing your library.

<hr>
<h1>
1: Fix our scoring functions (30 points)
</h1>
<p>
<div class=h1_cell>
<p>
As we defined our scoring functions in week 2, they all assume that the 4 cases have values greater than 0. But this week we will see that is not always the case. And hence you will get an error like this if you try using the scoring functions as is:
  <pre>
  KeyError: 'false_negative'
  </pre>
<p>
What this means is that there were no false negatives found so the key is also not found. You would find the problem in a statement like this:
  <pre>
  fn = cases['false_negative']  #key does not exist because no false negatives
  </pre>
  Your task is to fix up the 3 scoring functions so they handle cases like this. And while you are at it, make sure you never divide by 0. Once you have them working, you can add them to your library for this week. I give you some test cases to code against.
  <p>
 Important: do not import your library before (a) you have all functions below working correctly, and (b) you have replaced the old functions in your library with the new ones.
</div>

def accuracy(cases):
    tp = cases['true_positive'] if 'true_positive' in cases else 0
    tn = cases['true_negative'] if 'true_negative' in cases else 0
    fp = cases['false_positive'] if 'false_positive' in cases else 0
    fn = cases['false_negative'] if 'false_negative' in cases else 0
    return (tp + tn) / (tp+tn+fp+fn) if (tp+tn+fp+fn) != 0 else 0

#test

print(('should be 0', accuracy(pd.Series({}))))
print(('should be 0.3', accuracy(pd.Series({'true_positive':1, 'true_negative':2, 'false_positive':3, 'false_negative':4}))))
print(('should be 0.2', accuracy(pd.Series({'true_negative':2, 'false_positive':3, 'false_negative':4}))))

def f1(cases):
    #the heart of the matrix
    tp = cases['true_positive'] if 'true_positive' in cases else 0
    tn = cases['true_negative'] if 'true_negative' in cases else 0
    fp = cases['false_positive'] if 'false_positive' in cases else 0
    fn = cases['false_negative'] if 'false_negative' in cases else 0
    
    #other measures we can derive
    recall = 1.0*tp / (tp+fn) if (tp+fn) != 0 else 0 # positive correct divided by total positive in the table
    precision = 1.0*tp / (tp+fp) if (tp+fp) != 0 else 0 # positive correct divided by all positive predictions made
    
    #now for the one we want
    term1 = 1 / recall if recall != 0 else 0
    term2 = 1 / precision if precision != 0 else 0
    f1 = 2 / (term1 + term2) if term1 + term2 != 0 else 0
    
    return f1

#test

print(('should be 0', f1(pd.Series({}))))
print(('should be 0.2', f1(pd.Series({'true_positive':1, 'true_negative':2, 'false_positive':3, 'false_negative':4}))))
print(('should be 0', f1(pd.Series({'true_negative':2, 'false_positive':3, 'false_negative':4}))))

def informedness(cases):
    tp = cases['true_positive'] if 'true_positive' in cases else 0
    tn = cases['true_negative'] if 'true_negative' in cases else 0
    fp = cases['false_positive'] if 'false_positive' in cases else 0
    fn = cases['false_negative'] if 'false_negative' in cases else 0
    recall = 1.0*tp / (tp+fn) if (tp+fn) != 0 else 0  # positive correct divided by total positive in the table
    specificty = 1.0*tn/(tn+fp) if (tn+fp) != 0 else 0 # negative correct divided by total negative in the table
    J = (recall + specificty) - 1
    return J

#test

print(('should be -1', informedness(pd.Series({}))))
print(('should be -0.39', informedness(pd.Series({'true_positive':1, 'true_negative':2, 'false_positive':3, 'false_negative':4}))))
print(('should be -0.6', informedness(pd.Series({'true_negative':2, 'false_positive':3, 'false_negative':4}))))

<h1>Ok, now update your library</h1>
When you have done that, you are ready to import.
<p>
  <b>Important note</b>: you may be tempted to skip the step of updating your library before importing. Why do that when you already have the new functions defined above? We are back to that static-scope thing. Your notebook context (namespace) can see your library functions but your library functions cannot see your notebook functions. The library has its own namespace. In summary, `produce_scores` will not see your functions above. Instead it will see the functions in your library. And if you don't update your library, you will be using your old, broken versions. And it does not matter if you import first then define new functions in the notebook. Your notebook functions still will not override your library functions. Whew. So go update your library!

!rm library_w19_week5.py

from google.colab import files
files.upload()

from library_w19_week5 import *

%who function

<hr>
<h1>
2: Explore depth parameter with 10-folding (10 points)
</h1>
<p>
<div class=h1_cell>
<p>
In the week 4 module you were asked to explore depths from 1 to 5 without any cross-validation. Go ahead and do it again but using K-folding with K=10.
<p>
First, define the columns to use. I do that for you below.
</div>

splitter_columns = [
        #ApplicantIncome
       'apin_low', 'apin_high', 'apin_average',
        #Property_Area
        'pa_Rural', 'pa_Semiurban','pa_Urban',
        #LoanAmount
        'lam_Low', 'lam_Average', 'lam_High',
        #Credit_History
        'ch_bad', 'ch_good']

<hr>
<h2>
Show your results below
</h2>
<p>
<div class=h1_cell>
<p>
Explore depths 1-5 with 10-folding. You can leave the gig cutoff as default value.
</div>

table10_1 = k_fold(loan_table, 10, 'Loan_Status', {'max-depth':1}, splitter_columns)
table10_1

table10_2 = k_fold(loan_table, 10, 'Loan_Status', {'max-depth':2}, splitter_columns)
table10_2

table10_3 = k_fold(loan_table, 10, 'Loan_Status', {'max-depth':3}, splitter_columns)
table10_3

table10_4 = k_fold(loan_table, 10, 'Loan_Status', {'max-depth':4}, splitter_columns)
table10_4

table10_5 = k_fold(loan_table, 10, 'Loan_Status', {'max-depth':5}, splitter_columns)
table10_5

<hr>
<h1>
What does it look like to you?
</h1>
<p>
<div class=h1_cell>
<p>
To me it looks like the stump is best! As we add depth I don't see an improvement. And it looks like results from depth 4 and 5 the same. What I would try next is the gig cutoff and bringing more columns into play (e.g., Gender, Married) by applying ohe to them.
</div>

<hr>
<h1>
4: Random slices (60 points)
</h1>
<p>
<div class=h1_cell>
<p>
I'd like you to try one more thing. The `k_fold` function takes slices sequentially. I'd like you to write a function `k_fold_random` that takes random slices *without replacement*. Without replacement means that each slice should have a unique set of rows, i.e., no slices share a row
<p>
Each slice should be roughly 1/k percent of the entire table.
<p>
To solve this problem, I found it easiest to play around with loops in a notebook cell before trying to replace code in `k_fold`. My target was the equivalent of this but now with random slices:
<pre>
<code>
    total_len = len(table.index)
    slice_size = int(total_len/(1.0`*`k))
    slices = []

    #generate the slices
    for i in range(k-1):
        a_slice =  table[i`*`slice_size:(i+1)`*`slice_size]
        slices.append( a_slice )
    slices.append( table[(k-1)`*`slice_size:] )
    
    verify_unique(slices)  # I ask you to define this debugging function below
</code>
</pre>
<p>
I copied this code into a cell and then started to play around with various ideas until I got one I liked.
<p>
</div>

<h2>Help with debugging</h2>
I'll define a debugging function that will take a list of slices and do a pairwise comparison. For each pair, it will print the set intersection between the 2 slices. I give you a few test cases to try.

#Determine if slices are mutually exclusive
def verify_unique(slices):
    print(('total length all slices', sum([len(s) for s in slices])))
    for i, a_slice in enumerate(slices[:-1]):
        a_set = set(a_slice.index)
        for j, b_slice in enumerate(slices[i+1:]):
            b_set = set(b_slice.index)
            int_set = a_set.intersection(b_set)  # should be empty set as result
            print((i,j+i+1,int_set))
    return None

#test

verify_unique([loan_table[0:2], loan_table[2:4], loan_table[4:5]])  #looks good - enpty set for each pair

#test

verify_unique([loan_table[0:2], loan_table[1:4], loan_table[1:5]])  #not so good, finds intersection between all 3 pairs

# here is sequential slice code from k_fold if you want to use it as base.
# modify it to produce slices with random rows in each slice.
import random

table = loan_table
k = 10

total_len = len(table.index)
split_size = int(total_len/(1.0*k))
slices = []
available_rows = [i for i in range(total_len)] # we will pick a random row from here

#generate the slices
for i in range(k):
  a_slice = pd.DataFrame(columns=table.columns)
  
  for j in range(split_size):
    rand_idx = random.randint(0, len(available_rows) - 1)
    idx = available_rows[rand_idx] # which row we want to grab
    a_slice = a_slice.append(table[idx:idx+1], ignore_index=False) # add random row to current slice
    available_rows.remove(idx) # remove that row from the available rows
  
  slices.append( a_slice )

for remaining_idx in available_rows: # add the remaining rows to our final slice
  slices[len(slices) - 1] = slices[len(slices) - 1].append(table[remaining_idx:remaining_idx+1], ignore_index=False)
  
verify_unique(slices)  # should see 614 length and empty sets all the way down

<h2>
Plug your code in
</h2>
<p>
<div class=h1_cell>
<p>
Once you are generating slices randomly, you will have k test slices. You can go ahead and place your code into `k_fold_random`. You should now be good to go.
<p>
One problem with doing random selections is that we will get different results every time we run our code. We will see a way to get repeatable results in next module. For now, you may find that you are not matching my fold results.
</div>

def k_fold_random(table, k, target, hypers, candidate_columns):
  
    #set up the table where we will record fold results
    result_columns = ['name',  'accuracy', 'f1', 'informedness']
    k_fold_results_table = pd.DataFrame(columns=result_columns)
    
    #generate the slices
    total_len = len(table.index)
    split_size = int(total_len/(1.0*k))
    slices = []
    available_rows = [i for i in range(total_len)] # we will pick a random row from here

    for i in range(k):
      a_slice = pd.DataFrame(columns=table.columns)

      for j in range(split_size):
        rand_idx = random.randint(0, len(available_rows) - 1)
        idx = available_rows[rand_idx] # which row we want to grab
        a_slice = a_slice.append(table[idx:idx+1], ignore_index=False) # add random row to current slice
        available_rows.remove(idx) # remove that row from the available rows

      slices.append( a_slice )

    for remaining_idx in available_rows: # add the remaining rows to our final slice
      slices[len(slices) - 1] = slices[len(slices) - 1].append(table[remaining_idx:remaining_idx+1], ignore_index=False)
      
    verify_unique(slices)  # leaving this in here to verify it is working
    
    #generate test results
    all_scores = []  #keep track of all k results
    for i in range(k):
        test_table = slices[i]
        train_table = compute_training(slices, i)
        fold_tree = build_tree_iter(train_table, candidate_columns, target, hypers)  # train
        scores = produce_scores(test_table, fold_tree, target)  # test
        results_row = {'name': 'fold_'+str(i), 'accuracy': scores[0], 'f1': scores[1], 'informedness': scores[2]}
        k_fold_results_table = k_fold_results_table.append(results_row,ignore_index=True)
        all_scores.append(scores)
    
    #compute average of all folds
    avg_scores = tuple(reduce(lambda total, triple: np.add(triple, total), all_scores)/k)
    results_row = {'name': 'average', 'accuracy': avg_scores[0], 'f1': avg_scores[1], 'informedness': avg_scores[2]}
    k_fold_results_table = k_fold_results_table.append(results_row,ignore_index=True)
    
    #note that I add the meta comment as last step to avoid it being wiped out
    k_fold_results_table.meta = SimpleNamespace()
    k_fold_results_table.meta.hypers  = hypers # adds comment to remind me of hyper params used
    
    return k_fold_results_table


k_fold_random(loan_table, 10, 'Loan_Status', {'max-depth': 1}, splitter_columns)

What we got with prior sequential split at depth 1: 
<pre>
10	average	0.809206	0.876388	0.406950
</pre>

k_fold_random(loan_table, 10, 'Loan_Status', {'max-depth': 2}, splitter_columns)

What we got in sequential at depth 2:
<pre>
10	average	0.807566	0.875392	0.402404
</pre>
Let's try one more.

k_fold_random(loan_table, 10, 'Loan_Status', {'max-depth': 3}, splitter_columns)

What we got in sequential at depth 3:
<pre>
10	average	0.804288	0.872774	0.397404
</pre>
Random looks very close to sequential. With this small a dataset, I could see that happening. In general, random is prefered. With sequential you might be picking up some funny ordering of the rows. Maybe they listed the best applicants first.

<hr>
<h1>Did not change loan table</h1>
<div class=h1_cell>
<p>
No need to write it out.
</div>

<hr>
<h1>We did change library</h1>
<div class=h1_cell>
<p>
We added `verify_unique` and `k_fold_random`. Add them to your library and call new file `library_w19_5b.py`. I added the `b` to differentiate it from functions we added in intro notebook and top of this one.
</div>
