==> 399 - Data Science/6_random_forests.ipynb <==
<h1>
<center>
Module 6: Random Forests
</center>
</h1>
<div class=h1_cell>

You will be working with the loan table again.

</div>

import pandas as pd

from google.colab import drive
drive.mount('/content/gdrive')

with open('/content/gdrive/My Drive/class_tables/loan_table_week4.csv', 'r') as f:
  loan_table = pd.read_csv(f)

!rm library_w19_week6.py

from google.colab import files
files.upload()

from library_w19_week6 import *

%who function

pd.set_option('display.max_columns', None)
loan_table.head()

loan_table.columns.values

<hr>
<h1>
1: Explore forest options (20)
</h1>
<p>
<div class=h1_cell>
<p>
Check out the results you get from forests of size 5, 11, 17.
<p>
First, define the columns to use. I do that for you below.
</div>

splitter_columns = [
        #Dependents
        'dep_0', 'dep_1', 'dep_2', 'dep_3+',
        #ApplicantIncome
       'apin_low', 'apin_high', 'apin_average',
        #Property_Area
        'pa_Rural', 'pa_Semiurban','pa_Urban',
        #LoanAmount
        'lam_Low', 'lam_Average', 'lam_High',
        #Credit_History
        'ch_bad', 'ch_good']

<hr>
<h2>
Set seeds so get consistent results
</h2>
<p>
<div class=h1_cell>
<p>

</div>

import numpy as np
import random

rng = np.random.RandomState(42)  #Will pass as arg to pandas sample method
random.seed(2000)

forest1 = forest_builder(loan_table, splitter_columns, 'Loan_Status', hypers={'random-state':rng})
len(forest1)

forest_scores(loan_table, forest1, 'Loan_Status')

forest2 = forest_builder(loan_table, splitter_columns, 'Loan_Status', hypers={'total-trees':11, 'random-state':rng})
len(forest2)

forest_scores(loan_table, forest2, 'Loan_Status')

forest3 = forest_builder(loan_table, splitter_columns, 'Loan_Status', hypers={'total-trees':17, 'random-state':rng})
len(forest3)

forest_scores(loan_table, forest3, 'Loan_Status')

<hr>
<h1>
2: Implement Out of Bag testing (80)
</h1>
<p>
<div class=h1_cell>
<p>
Last module we looked at the use of K-Folding as a means to test our trees. Random Forests give us an alternative by using out of bag testing. Using notes from the content notebook this week, find a way to do prediction using the oob idea. As reminder, the set union of all the oob lists in a forest make up the testing set. If there is a row in loan_table that is not in any oob list, that row should be omitted from the test set. Further, a tree only gets to vote on a specific row if that row is in the tree's oob list.
  <p>
  I am going to leave it to you to come up with an algorithm for doing oob testing. If you get totally stuck, I can supply hints. For grading I am looking to make sure you only use oob rows for testing and that each individual tree only votes on rows in its own oob list.
    <p>
      It is worthwhile solving this problem given something like it will likely be on next midterm.
</div>

def get_oob_table(original_table, forest):
  # get union of all out of bag rows
  oob_union = []
  for tree in forest:
    oob_union = list(set(oob_union) | set(tree['oob']))

  # create new table using unioned oob rows
  testing_table = original_table.loc[oob_union]
  testing_table = testing_table.reset_index()
  
  return testing_table

def oob_forest_scores(table, forest, target):
  scratch_table = pd.DataFrame(columns=['prediction', 'actual'])
  
  # will only add a vote if a tree is supposed to vote on 'row'
  def oob_vote_taker(row, forest):
    votes = {0:0, 1:0}
    for tree in forest:
      if row['index'] not in tree['oob']: continue # check if its in the tree's oob list
      prediction = tree_predictor(row, tree)
      votes[prediction] += 1
    winner = 1 if votes[1]>votes[0] else 0  #ties go to 0
    return winner
  
  scratch_table['prediction'] = table.apply(lambda row: oob_vote_taker(row, forest), axis=1)
  scratch_table['actual'] = table[target]  # just copy the target column
  cases = scratch_table.apply(lambda row: predictor_case(row, pred='prediction', target='actual'), axis=1)
  vc = cases.value_counts()
  return [accuracy(vc), f1(vc), informedness(vc)]

<h2>Check your oob testing against my results</h2>

If you used the random seeds to build your trees, your results should be the same as mine. No randomness during oob testing.

#whole table from above: [0.8094462540716613, 0.8764519535374868, 0.4104956556082149]

testing_table_1 = get_oob_table(loan_table, forest1)
oob_forest_scores(testing_table_1, forest1, 'Loan_Status')

#from above: [0.7785016286644951, 0.8597938144329897, 0.3058599921011058]

testing_table_2 = get_oob_table(loan_table, forest2)
oob_forest_scores(testing_table_2, forest2, 'Loan_Status')

#from above: [0.8094462540716613, 0.8764519535374868, 0.4104956556082149]

testing_table_3 = get_oob_table(loan_table, forest3)
oob_forest_scores(testing_table_3, forest3, 'Loan_Status')

<h2>Not a lot of change</h2>

Using oob testing did not affect scores much. I think we would need to work with bigger tables, e.g., the 25K shelter table, to see a difference.

<hr>
<h1>Write it out</h1>
<div class=h1_cell>

Did not change table but we did define new functions. Add them to your library as `!rm library_w19_week6b.py`. I added the `b` to designate functions from assignment portion of module.
</div>
