==> 399 - Data Science/7_KNN.ipynb <==
<h1>
<center>
Module 8: K-NN
</center>
</h1>
<div class=h1_cell>
<p>
I kind of like K Nearest Neighbor (K-NN). It is a version of crowd-sourcing like random forests. The general idea is that you choose K to be how many "consultants" you take votes from. With random forests, the consultants were decision trees. In K-NN, they are the other rows in the table. So let's say you are trying to predict what value to give for a row i. What you do is check all the other rows (excluding i) to get their distance from i. You choose the K closest. You take the predictions from each of those K. Majority wins and is the prediction you use for row i.
<p>
Notice anything interesting? We are not doing any training! We do have to have labeled data to get the K consultants. But no training step involved. Wow.
<p>
However, there is always a catch. Think about what we are doing. For every row, we are calculating distance to every other row. Feels `O(n)` to me. And remember, the shelter data we looked at had 26K rows! What is cost of decision tree or random forest to make a prediction? Worst case it is the depth of the tree, i.e., it is constant speed. If we need to make real-time predictions then `O(n)` might not be good enough.
</div>
<p>
  And btw, when testing K-NN, our cost will balloon to `O(n**2)`.

import pandas as pd

from google.colab import drive
drive.mount('/content/gdrive')

with open('/content/gdrive/My Drive/class_tables/loan_table_week4.csv', 'r') as f:
  loan_table = pd.read_csv(f)

!rm library_w19_week6b.py

from google.colab import files
files.upload()

from library_w19_week6b import *

%who function

pd.set_option('display.max_columns', None)
loan_table.head()

loan_table.describe()

<h1>You are mostly on your own</h1>

I am going to give you the freedom to implement the K-NN components. My only constraints are that you implement these functions:
<pre>
def euclidean_distance(vector1, vector2):  #euclidean distance between 2 vectors

def knn(row_index, table, k, columns, target):  #provide prediction for row given K-NN algorithm. Only use values in columns.

def knn_tester(table, k, columns, target): #produce predictions for every row in table according to columns. Return accuracy score. Danger: O(n**2).
</pre>


I am calling these "splitter" columns but that is a misnomer. Splitter has to do with decision trees. Should rename them at some point.

splitter_columns = [
        #Dependents
        'dep_0', 'dep_1', 'dep_2', 'dep_3+',
        #ApplicantIncome
       'ApplicantIncome', 'CoapplicantIncome',
        #Property_Area
        'pa_Rural', 'pa_Semiurban','pa_Urban',
        #LoanAmount
        'filled_lam',
        #Credit_History
        'ch_bad', 'ch_good']

target = 'Loan_Status'

def euclidean_distance(vector1, vector2):
  if len(vector1) != len(vector2): 
    raise ValueError('Euclidean distance could not be calculated: vectors must be same length')
  
  dist = 0
  for i in range(len(vector1)):
    dist += (vector1[i] - vector2[i])**2
  
  dist = dist**(1/2)
  return dist

def knn(row_index, table, k, columns, target):
  row = table.loc[row_index, columns] # get target row
  row_list = list(row)
  
  # get all rows except for row_index
  other_rows = table.loc[table.index != row_index, columns]
  
  # calculate distances for each row
  distances = []
  for idx, other_row in other_rows.iterrows():
    distances.append([idx, euclidean_distance(row_list, list(other_row))])
  
  votes = {0:0, 1:0}
  distances = sorted(distances, key = lambda tup: tup[1])
  for i in range(k): # for each closest row
    curr_row = table.loc[distances[i][0]] 
    votes[curr_row[target]] += 1 # add its 'prediction' to the votes
  
  return 1 if votes[1] > votes[0] else 0

def knn_tester(table, k, columns, target):
  scratch_table = pd.DataFrame(columns=['prediction', 'actual'])
  scratch_table['prediction'] = table.apply(lambda row: knn(row.name, table, k, columns, target), axis = 1)
  scratch_table['actual'] = table[target]
  cases = scratch_table.apply(lambda row: predictor_case(row, pred='prediction', target='actual'), axis=1)
  vc = cases.value_counts()
  return accuracy(vc)

<h2>Match my results</h2>

Start with single predictions.

# test with row 1, k=5
import time

start = time.time()
    
prediction_row0 = knn(1, loan_table, 5, splitter_columns, target)        
print('prediction: ' + str(prediction_row0))  #actual is 0
end = time.time()
print('elapsed time: '+ str(end - start))  # in seconds

# fickas' output
# prediction: 1
# elapsed time: 0.19412565231323242

# my output
# prediction: 1
# elapsed time: 0.05439352989196777

#test with row 1, k=11

import time

start = time.time()
    
prediction_row0 = knn(1, loan_table, 11, splitter_columns, target)        
print('prediction: ' + str(prediction_row0))
end = time.time()
print('elapsed time: '+ str(end - start))  # in seconds

# fickas' output
# prediction: 1
# elapsed time: 0.18313097953796387

# my output
# prediction: 1
# elapsed time: 0.06505417823791504

<h2>Now to testing</h2>

knn_tester should use knn to make predictions for every row. Check predictions against target and compute and return accuracy.
<p>
All of my functions are nothing fancy. And it costs me about 2 minutes per run.

#k=5 (about 2 minutes on colab)

start = time.time()
    
the_accuracy = knn_tester(loan_table, 5, splitter_columns, target)        
print('accuracy: ' + str(the_accuracy))
end = time.time()
print('elapsed time: '+ str(end - start))  # in seconds

# fickas' output
# accuracy: 0.6302931596091205
# elapsed time: 109.74176526069641

# my output
# accuracy: 0.6302931596091205
# elapsed time: 29.347878456115723

<h2>Not that great</h2>

Why so bad? Normally K-NN fails when there are too many features/columns in play. It relies on distances, and the more columns you add, the harder to discern distance differences. This is an example of the curse of dimensionality (jargon alert). It is kind of counter-intuitive: the more information (features/columns) you add, the worse you do. Here is an intro to the problem: http://www.visiondummy.com/2014/04/curse-dimensionality-affect-classification/. Let's reduce our columns and see if it makes any difference.


fewer_columns = [
        #ApplicantIncome
       'ApplicantIncome', 'CoapplicantIncome',
        #LoanAmount
        'filled_lam',
        #Credit_History
        'ch_bad', 'ch_good']

#k=5 (about 2 minutes on colab)

start = time.time()
    
the_accuracy = knn_tester(loan_table, 5, fewer_columns, target)        
print('accuracy: ' + str(the_accuracy))
end = time.time()
print('elapsed time: '+ str(end - start))  # in seconds

# fickas' output
# accuracy: 0.6302931596091205
# elapsed time: 83.48994851112366

# my output
# accuracy: 0.6302931596091205
# elapsed time: 21.349454641342163

Exactly the same. Let's try some different values for k.

#k = 11, full columns
start = time.time()
    
the_accuracy = knn_tester(loan_table, 11, splitter_columns, target)        
print('accuracy: ' + str(the_accuracy))
end = time.time()
print('elapsed time: '+ str(end - start))  # in seconds

# fickas' output
# accuracy: 0.6628664495114006
# elapsed time: 110.49167799949646

# my output
# accuracy: 0.6628664495114006
# elapsed time: 29.346335649490356

Got a bit of improvement. One more.

#k = 3, full columns
start = time.time()
    
the_accuracy = knn_tester(loan_table, 3, splitter_columns, target)        
print('accuracy: ' + str(the_accuracy))
end = time.time()
print('elapsed time: '+ str(end - start))  # in seconds

# fickas' output
# accuracy: 0.6042345276872965
# elapsed time: 108.1471917629242

# my output
# accuracy: 0.6042345276872965
# elapsed time: 27.696593761444092

Heading back down in accuracy. I'll stop here but will leave you with an extra credit assignment (optional).

<h2>Update your library</h2>

Name it `library_w19_week7.py`.

<h2>Extra Credit</h2>

Beat my testing times. You can use any of the pandas and numpy methods. As reminder, here were my times (approximately) using brute force approach:
<pre>
knn_tester(loan_table, 5, splitter_columns, target) : 107ish
knn_tester(loan_table, 5, fewer_columns, target) :  83ish
knn_tester(loan_table, 11, splitter_columns, target) : 108ish
knn_tester(loan_table, 3, splitter_columns, target) : 108ish
</pre>
<p>
  Here is a time I got by a fairly simple trick:
  <pre>
  knn_tester(loan_table, 3, splitter_columns, target) 
  accuracy: 0.6042345276872965
elapsed time: 4.287370681762695
</pre>
So from 108 to 4. Pretty dang good.
